Write tests for the clear all completed button
